Starting with Multiprocessing
Starting to create a data_array. The idea with this array is to create a space that will access the FlightLog_Waypoints.csv file. It will have a timer that will update the folder each time the time runs out. It will store it to the data array, then the data array will be passed to the tcp process and the gui processing.